(core) Move faceid_db into core/db.py
Move the pickle + jsonl data layer into the core package. faceid_db.py
becomes a compatibility shim that aliases itself to core.db in sys.modules,
so `import faceid_db` and `import core.db` resolve to the same module object.
This keeps attribute access and test monkeypatching coherent across both
names while the ~19 importers migrate.

Retarget test_faceid_db.py and core-internal imports (config, naming) to
core.db; add core.db to spec hiddenimports.
