+++ backend/core/config.py
@@ -20,7 +20,7 @@ from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-from faceid_db import ARCHIVE_DIR, ATTEMPT_SETTINGS_SIG, BASE_DIR, CONFIG_PATH, LOGGING_PATH
+from core.db import ARCHIVE_DIR, ATTEMPT_SETTINGS_SIG, BASE_DIR, CONFIG_PATH, LOGGING_PATH
 
 if TYPE_CHECKING:
     import numpy as np
